pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-classifier"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "sentence-transformers>=3.0.0",
    "datasets>=2.19.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "datasketch>=1.6.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.66.0",
    "accelerate>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_classifier"]

[tool.pytest.ini_options]
testpaths = ["tests"]
